* move component to registry directory
